Make the engine link and load under MSVC
Three Windows facts the GCC build never had to answer, found by the
Windows CI leg:

Jolt defaults to the static MSVC runtime while the rest of the build
uses the dynamic one, so every Jolt object failed to link into the
engine with LNK2038. Force its option off.

MSVC exports nothing from a DLL without a declspec or a module
definition file, where ELF and MinGW export every non-hidden symbol.
Let CMake generate the export table so consumers see one surface on
every toolchain.

Windows resolves a DLL from the executable's own directory and has no
rpath, so a test binary in its own subdirectory cannot find
sencha_engine.dll. Collect runtime artifacts in one directory there,
leaving the layout scripts and docs address unchanged elsewhere.

Also two MSVC diagnostics worth fixing rather than silencing: Registry
is a struct but four headers forward-declared it as a class (C4099),
and the shadow caster range test narrowed a double to a float
implicitly (C4244).

+++ engine/src/render/ShadowDepthPass.cpp
@@ -52,7 +52,7 @@ namespace
             const double closest = std::clamp(value,
                                               static_cast<double>(bounds.Min[axis]),
                                               static_cast<double>(bounds.Max[axis]));
-            delta[axis] = closest - value;
+            delta[axis] = static_cast<float>(closest - value);
         }
         const double range = static_cast<double>(sphere.W);
         return delta.SqrMagnitude() <= range * range;
